[ng] automake: new global variable '%clean_dirs'
Similar to the '%clean_files' variable, but it will allow us to declare
directories (rather than regular files) that are to be cleaned by the
four flavours of cleaning targets ('mostlyclean', 'clean', 'distclean',
'maintainer-clean').  This will be especially useful in later changes.

* automake.in (%clean_dirs): New global.
(initialize_per_input): Reset it.
* automake.in (handle_clean), lib/am/clean.am: Extend to clean the
directories registered in '%clean_dirs'.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>