[ng] cleanup: remove 'libtool.am'
It's so small and dumb that it's easier and cleaner to just inline
it in the automake script.

* lib/am/libtool.am: Delete.
* Makefile.am (dist_am_DATA): Remove it.
* automake.in (handle_libtool): Just add the list of all the '.libs'
directories to '%libtool_clean_directories', instead of processing the
now-removed 'libtool.am' to obtain the same effect.  Similarly, when
processing the Makefile in the top-level directory, add 'libtool' and
'config.lt' to '%clean_files'.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>