contrib: drop 'multilib' support
Its tests are hopelessly failing now, it is complex to debug (and I'm
too ignorant about its aims and its implementation to do so), and anyway
it should probably be rewritten to take better advantage of GNU make and
Automake-NG features.  So let's drop it for now, to avoid slowing down
the Automake-NG development.  Anyone interested in the 'multilib' will
send patches re-implementing it after out first Automake-NG beta is out.
Probably several months from now.

* contrib/multilib/README, contrib/multilib/config-ml.in,
* contrib/multilib/multi.m4, contrib/multilib/multilib.am,
* contrib/multilib/symlink-tree, t/help-multilib.sh,
* t/multlib.sh: Remove.
* Makefile.am (EXTRA_DIST): Adjust.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>