contrib: simple improvements to check-html
Related to automake bug#11287.

* contrib/t/check-html.am (.log.html): Do not look for $RST2HTML in
the environment, looking for $(RST2HTML) should be enough (also, the
pre-existing code was broken, because it single-quoted $RST2HTML).
Fix the loop-and-search implementation to be more similar to the
one in 'contrib/t/parallel-tests-html.sh'.  Prefer 'mv -f' over bare
'mv'.  Do not break the recipe in

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>