[ng] check: in recipes, don't use pipe when redirection suffices
Instead of the botched idiom:

    while read c; do echo $c; done <file | command

use the simpler and much better one:

    command <file

Not only the latter is simpler and marginally more efficient, but is also
better at catching possible unexpected I/O errors; while in the former,
the exit status of the while loop (which might have revealed such errors)
was lost in the pipeline.

* lib/am/parallel-tests.am ($(TEST_SUITE_LOG), recheck): Use the better
idiom.

Co-authored-by: Akim Demaille <akim@lrde.epita.fr>
Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>