automake: '--add-missing' is on by default
This commit modifies the automake script will always act as if the
'--add-missing' option were give, and simplifies the documentation
accordingly.

Future commits will take care of cleaning up the code accordingly.

The new semantics introduced with this change should have ideally
been the default in automake mainline as well, but they are not,
likely for historical reasons (as is the case for too many Automake
warts).

The new semantics also prepare the way for an exciting planning
change: the distribution of (at first only few) makefile fragments
to be included at make runtime in the Automake-generated Makefiles,
instead of having their contents copied over and over in said
Makefiles.

* automake.in ($add_missing): Define to '1' unconditionally.
(parse_arguments): The '-a' a.k.a. '--add-missing' option is
now a no-op.
* doc/automake-ng.texi: Adjust, in particular with the addition
of ...
(Auxiliary files): ... this new node.
* NG-NEWS: Update.
* t/add-missing.tap: Simplify: don't check that , if any required
auxiliary file is missing and automake is called without the '-a'
or '--add-missing' option, it complains and error out; this is not
the case anymore.  Also, don't call automake with the '-a' or
'--add-missing' option; there's no need to now.
* t/dist-auxfile.sh: Adjust.
* t/license2.sh: Likewise.
* t/per-target-flags.sh: Likewise.
* t/python.sh: Likewise.
* t/missing-auxfile-stops-makefiles-creation.sh: Remove as obsolete.
* t/mdate.sh: Likewise.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>