[ng] vars: remove variable $(am__sh_e_setup)
The comments to that variable claimed that "set +e" in the shell was
an XSI extension, but the Automake testsuite since at least 1.11 (which
was run several times with some weak, non-POSIX vendor /bin/sh) used it
in the test script, with not problem whatsoever.  So we don't actually
need the safety net provided by $(am__sh_e_setup): we can simply use
'set +e' in the relevant recipes.

* lib/am/header-vars.am (am__sh_e_setup): Remove.
* lib/am/parallel-tests.am: Remove usages; simply inline 'set +e' instead.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>