[ng] automake: move processing of config-header rules in its own function
This is a pure refactoring with no semantic changes intended.  It will
only be useful for future changes.

* automake.in (handle_configure): Move definition of $(AM_CONF_HEADERS)
from here ...
(handle_config_headers): ... to this new function, which will also ensure
that the $(AM_CONFIG_HEADERS) will always be defined (albeit possibly to
an empty value).
(generate_makefile): Call the new variable.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>